requirements.txt
mysql-connector-python==8.0.18
Flask==1.0.2
requests==2.19.1
gunicorn